change for ksy textureview
