+++ KsyunPlayerDemo/app/src/main/java/com/ksyun/player/demo/activity/NetMediaActivty.java
@@ -2,6 +2,7 @@ package com.ksyun.player.demo.activity;
 
 import android.content.Context;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.database.Cursor;
 import android.os.Bundle;
 import android.support.v7.app.ActionBar;
@@ -16,8 +17,9 @@ import android.widget.LinearLayout;
 import android.widget.ListView;
 import android.widget.Toast;
 
-import com.ksyun.player.demo.model.NetDbAdapter;
 import com.ksyun.player.demo.R;
+import com.ksyun.player.demo.model.NetDbAdapter;
+import com.ksyun.player.demo.util.Settings;
 
 import java.util.ArrayList;
 
@@ -34,6 +36,7 @@ public class NetMediaActivty extends AppCompatActivity implements View.OnClickLi
 
     private Cursor cursor;
     private NetDbAdapter NetDb;
+    private SharedPreferences settings;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -71,10 +74,21 @@ public class NetMediaActivty extends AppCompatActivity implements View.OnClickLi
                     NetDb.createDate(path);
                 }
                 NetDb.close();
-                Intent intent = new Intent(NetMediaActivty.this,VideoPlayerActivity.class);
-                intent.putExtra("path",path);
-                startActivity(intent);
 
+                String chooseview;
+                settings = getSharedPreferences("SETTINGS",Context.MODE_PRIVATE);
+                chooseview = settings.getString("choose_view","undefind");
+
+               if(chooseview.equals(Settings.USEKSYTEXTURE)){
+                    Intent intent = new Intent(NetMediaActivty.this, TextureVideoActivity.class);
+                    intent.putExtra("path",path);
+                    startActivity(intent);
+                }else{
+                    Intent intent = new Intent(NetMediaActivty.this,SurfaceActivity.class);
+                    intent.putExtra("path",path);
+                    startActivity(intent);
+
+                }
             }
         });
         setActionBarLayout(R.layout.net_actionbar,this);
