+++ KsyunPlayerDemo/app/src/main/java/com/ksyun/player/demo/activity/SettingActivity.java
@@ -2,8 +2,8 @@ package com.ksyun.player.demo.activity;
 
 import android.content.Context;
 import android.content.SharedPreferences;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
 import android.widget.CompoundButton;
 import android.widget.RadioButton;
 import android.widget.RadioGroup;
@@ -22,7 +22,7 @@ public class SettingActivity extends AppCompatActivity implements RadioGroup.OnC
     private SharedPreferences settings ;
     private SharedPreferences.Editor editor;
     private RadioButton radiosurface;
-    private RadioButton radiotexture;
+    private RadioButton radioksytexture;
     private RadioButton radiosoft;
     private RadioButton radiohard;
     private Switch debugswitch;
@@ -41,7 +41,6 @@ public class SettingActivity extends AppCompatActivity implements RadioGroup.OnC
         String choosedecode = settings.getString("choose_decode","信息为空");
         String choosedebug = settings.getString("choose_debug","信息为空");
 
-
         mChooseSurface = (RadioGroup) findViewById(R.id.choose_surface);
         mChooseCodec = (RadioGroup) findViewById(R.id.choose_codec);
 
@@ -56,10 +55,10 @@ public class SettingActivity extends AppCompatActivity implements RadioGroup.OnC
             @Override
             public void onCheckedChanged(CompoundButton compoundButton, boolean isChecked) {
                 if (isChecked) {
-                    editor.putString("choose_debug",Settings.DEBUGON);
+                    editor.putString("choose_debug", Settings.DEBUGON);
                     Toast.makeText(SettingActivity.this, "Debug被打开", Toast.LENGTH_SHORT).show();
                 } else {
-                    editor.putString("choose_debug",Settings.DEBUGOFF);
+                    editor.putString("choose_debug", Settings.DEBUGOFF);
                     Toast.makeText(SettingActivity.this, "Debug被关闭", Toast.LENGTH_SHORT).show();
                 }
                 editor.commit();
@@ -71,7 +70,7 @@ public class SettingActivity extends AppCompatActivity implements RadioGroup.OnC
         radiosoft = (RadioButton)findViewById(R.id.use_sw);
         radiohard = (RadioButton)findViewById(R.id.use_hw);
         radiosurface = (RadioButton)findViewById(R.id.use_surfaceview);
-        radiotexture = (RadioButton)findViewById(R.id.use_textureview);
+        radioksytexture = (RadioButton)findViewById(R.id.use_ksytextureview);
 
         switch (choosedecode){
             case Settings.USEHARD:
@@ -82,19 +81,19 @@ public class SettingActivity extends AppCompatActivity implements RadioGroup.OnC
                 break;
             default:
                 mChooseCodec.check(radiosoft.getId());
-                editor.putString("choose_decode",Settings.USESOFT);
+                editor.putString("choose_decode", Settings.USESOFT);
                 break;
         }
         switch (chooseview){
             case Settings.USESUFACE:
                 mChooseSurface.check(radiosurface.getId());
                 break;
-            case Settings.USETEXTURE:
-                mChooseSurface.check(radiotexture.getId());
+            case Settings.USEKSYTEXTURE:
+                mChooseSurface.check(radioksytexture.getId());
                 break;
             default:
                 mChooseSurface.check(radiosurface.getId());
-                editor.putString("choose_view",Settings.USESUFACE);
+                editor.putString("choose_view", Settings.USESUFACE);
                 break;
         }
         switch (choosedebug){
@@ -106,11 +105,10 @@ public class SettingActivity extends AppCompatActivity implements RadioGroup.OnC
                 break;
             default:
                 debugswitch.setChecked(false);
-                editor.putString("choose_debug",Settings.DEBUGOFF);
+                editor.putString("choose_debug", Settings.DEBUGOFF);
                 break;
         }
         editor.commit();
-
 
     }
 
@@ -119,21 +117,21 @@ public class SettingActivity extends AppCompatActivity implements RadioGroup.OnC
     public void onCheckedChanged(RadioGroup radioGroup, int i) {
         switch(i){
             case R.id.use_surfaceview:
-                editor.putString("choose_view",Settings.USESUFACE);
+                editor.putString("choose_view", Settings.USESUFACE);
                 break;
-            case R.id.use_textureview:
-                editor.putString("choose_view",Settings.USETEXTURE);
+            case R.id.use_ksytextureview:
+                editor.putString("choose_view", Settings.USEKSYTEXTURE);
                 break;
             case R.id.use_hw:
-                mChooseSurface.check(radiotexture.getId());
+                mChooseSurface.check(radioksytexture.getId());
                 //radiosurface.setEnabled(false);
-                Toast.makeText(SettingActivity.this, "硬解请使用TextureView", Toast.LENGTH_SHORT).show();
-                editor.putString("choose_decode",Settings.USEHARD);
-                editor.putString("choose_view",Settings.USETEXTURE);
+                Toast.makeText(SettingActivity.this, "硬解请使用KSYTextureView", Toast.LENGTH_SHORT).show();
+                editor.putString("choose_decode", Settings.USEHARD);
+                editor.putString("choose_view", Settings.USEKSYTEXTURE);
                 break;
             case  R.id.use_sw:
                 //radiosurface.setEnabled(true);
-                editor.putString("choose_decode",Settings.USESOFT);
+                editor.putString("choose_decode", Settings.USESOFT);
                 break;
             default:
                 break;
