+++ KsyunPlayerDemo/app/src/main/java/com/ksyun/player/demo/util/Settings.java
@@ -5,7 +5,7 @@ package com.ksyun.player.demo.util;
  */
 public class Settings {
     public static final String USESUFACE="usesufaceview";
-    public static final String USETEXTURE="usetextureview";
+    public static final String USEKSYTEXTURE="useksytextureview";
     public static final String USEHARD="useharddecode";
     public static final String USESOFT="usesoftdecode";
     public static final String DEBUGON="debugon";
