+++ KsyunPlayerDemo/app/src/main/res/layout/activity_setting.xml
@@ -62,10 +62,11 @@
                 android:layout_height="wrap_content"
                 android:checked="true"
                 android:text="@string/set_surface" />
-            <RadioButton android:id="@+id/use_textureview"
+
+            <RadioButton android:id="@+id/use_ksytextureview"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="@string/set_texture" />
+                android:text="@string/set_ksytexture" />
 
         </RadioGroup>
